requirements.txt
PyQt5>=5.15,<6
openai>=1.75.0
mss>=10.0.0
Pillow>=10.4.0
